use 0.0f style constants instead of static_cast<float> (0.0)
* data.cc, det.cc, op-fcm-fcs.cc, op-fcs-fcm.cc, op-fcs-fcs.cc,
op-fcs-fm.cc, op-fcs-fs.cc, op-fm-fcs.cc, op-fs-fcs.cc,
fCDiagMatrix.cc, lo-mappers.cc, mx-ops: Use 0.0f style constants
instead of static_cast<float> (0.0).

+++ libinterp/corefcn/data.cc
@@ -4668,7 +4668,7 @@ either @qcode{\"double\"} or @qcode{\"single\"}.\n\
                     {
                       int expon;
                       gnulib::frexpf (val, &expon);
-                      epsval(i) = std::pow (static_cast <float> (2.0),
+                      epsval(i) = std::pow (2.0f,
                                             static_cast <float> (expon - 24));
                     }
                 }
